// src/Front-End/reru/projects/evaluation360/src/app/utils/models/auto-evaluate-dto.ts
export class AutoEvaluateDto {
	accept: boolean;
	question1SelfEvaluationMark: number;
	question2SelfEvaluationMark: number;
	question3SelfEvaluationMark: number;
	question4SelfEvaluationMark: number;
	question5SelfEvaluationMark: number;
	question6SelfEvaluationMark: number;
	question7SelfEvaluationMark: number;
	question8SelfEvaluationMark: number;
	question9SelfEvaluationMark: number;
	question10SelfEvaluationMark: number;
	question11SelfEvaluationMark: number;
	question12SelfEvaluationMark: number;
	question13SelfEvaluationMark: number;
	question14SelfEvaluationMark: number;
	question15SelfEvaluationMark: number;
	evaluatedComments: string;
	evaluatedAcceptance: number;
	individualObjective1: any;
	individualObjective2: any;
	individualObjective3: any;
	lastQualifyCategory: any;
	lastQualifyCategoryOrderDate: any;
	lastQualifyCategoryOrderNumber: any;
	lastQualifyCategoryOrderSeries: any;
	lastYearMark: number;
	degree: number;
	partialEvaluationFromDate: any;
	partialEvaluationMark: number;
	partialEvaluationRate: any;
	partialEvaluationToDate: any;
	sanctionEndDate: any;
	sanctionStartDate: any;
	specializationCoursesInternal: any;
	specializationCoursesInternational: any;
	specializationCoursesOthers: any;
	treeYearsAgoMark: number;
	twoYearsAgoMark: number;

	needImprovement1: string;
	needImprovement2: string;
	needImprovement3: string;

	question1Comments: string;
	question2Comments: string;
	question3Comments: string;
	question4Comments: string;
	question5Comments: string;
	question6Comments: string;

	objective1Name: string;
	objective2Name: string;
	objective3Name: string;
	objective4Name: string;
	objective5Name: string;

	objective1Performance: string;
	objective2Performance: string;
	objective3Performance: string;
	objective4Performance: string;
	objective5Performance: string;

	objective1Complete: string;
	objective2Complete: string;
	objective3Complete: string;
	objective4Complete: string;
	objective5Complete: string;

	objective1Mark: number;
	objective2Mark: number;
	objective3Mark: number;
	objective4Mark: number;
	objective5Mark: number;

	objective1Comments: string;
	objective2Comments: string;
	objective3Comments: string;
	objective4Comments: string;
	objective5Comments: string;

	question1Mark: number;
	question2Mark: number;
	question3Mark: number;
	question4Mark: number;
	question5Mark: number;
	question6Mark: number;
}